@@ -1,39 +1,39 @@
 source 'https://rubygems.org'
-ruby '2.3.0'
+ruby '2.5.1'
 
 # Standard Rails gems
-gem 'rails', '4.2.5'
-gem 'sass-rails', '4.0.4'
-gem 'uglifier', '2.5.3'
-gem 'coffee-rails', '4.1.0'
-gem 'jquery-rails', '3.1.2'
-gem 'turbolinks', '2.5.2'
-gem 'jbuilder', '2.2.5'
-gem 'bcrypt', '3.1.9'
-gem 'octopress', '~> 3.0.0.rc'
+gem 'rails', '5.2.1'
+gem 'sass-rails'
+gem 'uglifier'
+gem 'coffee-rails'
+gem 'jquery-rails'
+gem 'turbolinks'
+gem 'jbuilder'
+gem 'bcrypt'
+gem 'octopress'
 gem 'pg'
 gem 'sidekiq'
 
 # Kaminari: https://github.com/amatsuda/kaminari
-gem 'kaminari', '0.16.1'
+gem 'kaminari'
 
 # Friendly_id: https://github.com/norman/friendly_id
-gem 'friendly_id', '5.0.4'
+gem 'friendly_id'
 
 # Font-awesome: https://github.com/FortAwesome/font-awesome-sass
-gem 'font-awesome-sass', '4.2.2'
+gem 'font-awesome-sass'
 
 # Bootstrap 3: https://github.com/twbs/bootstrap-sass
-gem 'bootstrap-sass', '3.3.1.0'
+gem 'bootstrap-sass'
 
 # Figaro: https://github.com/laserlemon/figaro
 group :development, :test do
-  gem 'figaro', '1.0.0'
+  gem 'figaro'
   gem 'launchy'
 end
 
 # Devise: https://github.com/plataformatec/devise
-gem 'devise', '3.4.1'
+gem 'devise'
 gem 'omniauth'
 gem 'omniauth-twitter'
 gem 'omniauth-facebook'
@@ -42,7 +42,7 @@ gem 'omniauth-google'
 gem 'omniauth-github'
 
 # Redcarpet: https://github.com/vmg/redcarpet
-gem 'redcarpet', '3.2.0'
+gem 'redcarpet'
 
 # Unicorn: http://unicorn.bogomips.org
 group :production do
