@@ -3,9 +3,9 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  before_filter :configure_permitted_parameters, if: :devise_controller?
-  before_filter :reject_locked!, if: :devise_controller?
-  before_filter :ensure_signup_complete, only: [:new, :create, :update, :destroy]
+  before_action :configure_permitted_parameters, if: :devise_controller?
+  before_action :reject_locked!, if: :devise_controller?
+  before_action :ensure_signup_complete, only: [:new, :create, :update, :destroy]
 
 
   # Devise permitted params
